+++ c/bam_access.h
@@ -1,6 +1,6 @@
 /*       LICENCE
 * PCAP - NGS reference implementations and helper code for the ICGC/TCGA Pan-Cancer Analysis Project
-* Copyright (C) 2014 ICGC PanCancer Project
+* Copyright (C) 2014-2016 ICGC PanCancer Project
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU General Public License
@@ -25,8 +25,8 @@
 #include <stdlib.h>
 #include "htslib/sam.h"
 #include "dbg.h"
-
 #include "khash.h"
+
 KHASH_MAP_INIT_INT(ins,uint64_t)
 //KHASH_INIT2(ins,, khint32_t, uint64_t, 1, kh_int_hash_func, kh_int_hash_equal)
 
@@ -52,10 +52,10 @@ typedef struct{
   char *sample;
 } rg_info_t;
 
-rg_info_t **parse_header(bam_hdr_t *head, int *grps_size, stats_rd_t ****grp_stats);
+rg_info_t **bam_access_parse_header(bam_hdr_t *head, int *grps_size, stats_rd_t ****grp_stats);
 
-int process_reads(htsFile *input, bam_hdr_t *head, rg_info_t **grps, int grps_size, stats_rd_t ****grp_stats, int rna);
+int bam_access_process_reads(htsFile *input, bam_hdr_t *head, rg_info_t **grps, int grps_size, stats_rd_t ****grp_stats, int rna);
 
-uint64_t get_mapped_base_count_from_cigar(bam1_t *b);
+uint64_t bam_access_get_mapped_base_count_from_cigar(bam1_t *b);
 
 #endif