+++ c/bam_stats_output.h
@@ -1,0 +1,27 @@
+/*       LICENCE
+* PCAP - NGS reference implementations and helper code for the ICGC/TCGA Pan-Cancer Analysis Project
+* Copyright (C) 2014 ICGC PanCancer Project
+*
+* This program is free software; you can redistribute it and/or
+* modify it under the terms of the GNU General Public License
+* as published by the Free Software Foundation; either version 2
+* of the License, or (at your option) any later version.
+*
+* This program is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+* GNU General Public License for more details.
+*
+* You should have received a copy of the GNU General Public License
+* along with this program; if not see:
+*   http://www.gnu.org/licenses/gpl-2.0.html
+*/
+
+#ifndef __bam_stats_output_h__
+#define __bam_stats_output_h__
+
+#include "bam_access.h"
+
+int bam_stats_output_print_results(rg_info_t **grps,int grps_size,stats_rd_t*** grp_stats,char *input_file,char *output_file);
+
+#endif